requirements.txt
pandas
transformers[torch]
datasets
evaluate
diffusers
torch 
torchvision 
torchaudio
pillow
matplotlib
imagehash